feat: extract AppServerClient and JSON-RPC primitives into client module
Move the full JSON-RPC client implementation (AppServerClient, message
formatting/parsing, and direct spawn connection) from protocol.ts into
a new client.ts. The connect function is renamed to connectDirect to
distinguish it from future broker-based connection strategies.

protocol.ts becomes a thin re-export shim that maps connectDirect back
to connect, preserving backward compatibility for all existing consumers.
